+++ src/PostgREST/AppState.hs
@@ -32,8 +32,8 @@ import qualified Hasql.Session as SQL
 
 import Control.AutoUpdate (defaultUpdateSettings, mkAutoUpdate,
                            updateAction)
-import Data.IORef         (IORef, atomicWriteIORef, newIORef,
-                           readIORef)
+import Data.IORef         (IORef, atomicModifyIORef, atomicWriteIORef,
+                           newIORef, readIORef)
 import Data.Time          (ZonedTime, defaultTimeLocale, formatTime,
                            getZonedTime)
 import Data.Time.Clock    (UTCTime, getCurrentTime)
@@ -46,8 +46,8 @@ import Protolude
 
 
 data AppState = AppState
-  -- | Database connection pool
-  { statePool         :: SQL.Pool
+  -- | Database connection pool; 'Nothing' indicates the pool has been destroyed
+  { statePool         :: IORef (Maybe SQL.Pool)
   -- | Database server version, will be updated by the connectionWorker
   , statePgVersion    :: IORef PgVersion
   -- | No schema cache at the start. Will be filled in by the connectionWorker
@@ -74,13 +74,14 @@ data AppState = AppState
 
 init :: AppConfig -> IO AppState
 init conf = do
-  newPool <- initPool conf
-  initWithPool newPool conf
+  pool <- initPool conf
+  initWithPool pool conf
 
 initWithPool :: SQL.Pool -> AppConfig -> IO AppState
-initWithPool newPool conf =
-  AppState newPool
-    <$> newIORef minimumPgVersion -- assume we're in a supported version when starting, this will be corrected on a later step
+initWithPool pool conf =
+  AppState
+    <$> newIORef (Just pool)
+    <*> newIORef minimumPgVersion -- assume we're in a supported version when starting, this will be corrected on a later step
     <*> newIORef Nothing
     <*> newIORef mempty
     <*> newEmptyMVar
@@ -93,23 +94,59 @@ initWithPool newPool conf =
     <*> newIORef 0
 
 destroy :: AppState -> IO ()
-destroy AppState{..} = SQL.release statePool
+destroy = destroyPool
 
 initPool :: AppConfig -> IO SQL.Pool
 initPool AppConfig{..} =
-  SQL.acquire (configDbPoolSize, configDbPoolTimeout, toUtf8 configDbUri)
+  SQL.acquire configDbPoolSize $ toUtf8 configDbUri
 
+-- | Run an action with a database connection.
+--
+-- This is a thin wrapper around 'SQL.use', which deals with a
+-- race on flushing the pool: The pool read via 'readIORef'
+-- might be released via 'flushPool' by the time we call 'SQL.use',
+-- so we catch that 'SQL.PoolIsReleasedUsageError' and retry.
+--
+-- If on the other hand the pool has been (permanently) destroyed,
+-- we re-use 'SQL.PoolIsReleasedUsageError' to convey this to the
+-- caller.
 usePool :: AppState -> SQL.Session a -> IO (Either SQL.UsageError a)
-usePool AppState{..} = SQL.use statePool
+usePool appState@AppState{..} session = do
+  pool <- readIORef statePool
+  case pool of
+    Nothing -> return $ Left SQL.PoolIsReleasedUsageError
+    Just p -> do
+      res <- SQL.use p session
+      case res of
+        Left SQL.PoolIsReleasedUsageError -> usePool appState session
+        _                                 -> return res
 
 -- | Flush the connection pool so that any future use of the pool will
 -- use connections freshly established after this call.
 --
--- FIXME: #2401 Connections that are in-use during the call to flushPool
--- will currently be returned to the pool and reused afterwards, in
--- conflict with the intention.
+-- Does nothing if the pool has been destroyed.
 flushPool :: AppState -> IO ()
-flushPool AppState{..} = SQL.release statePool
+flushPool appState@AppState{..} = do
+  conf <- getConfig appState
+
+  -- create a new pool; this is discarded if the pool has been destroyed,
+  -- but it's not a huge waste
+  newPool <- initPool conf
+  oldPool <- atomicModifyIORef statePool $ \old ->
+    case old of
+      Nothing -> (Nothing, old)
+      _       -> (Just newPool, old)
+  case oldPool of
+    Nothing  -> return ()
+    Just old -> SQL.release old
+
+-- | Destroy the pool on shutdown.
+destroyPool :: AppState -> IO ()
+destroyPool AppState{..} = do
+  pool <- readIORef statePool
+  case pool of
+    Nothing -> return ()
+    Just p  -> SQL.release p
 
 getPgVersion :: AppState -> IO PgVersion
 getPgVersion = readIORef . statePgVersion